// src/app/models/marketing-management/post/post-model.ts
export class PostModel {
   _id: string;
   title: string;
   content: string;
   status: number;
   thumbnailImage: string;
   shortDescription: string;
   description: string;
   servicesId: string;
   publishDate: string;
   updatedDate: string;
   createdDate: string;
   updatedUserId: string;
   createdUserId: string;
}